backend/pyproject.toml
[project]
name = "studybuddy-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    # Avoid fastapi[standard] on Render: it pulls rignore (Rust/maturin) which
    # fails with read-only cargo cache. FastAPI + uvicorn[standard] is enough.
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
    "pyjwt[crypto]>=2.10.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.7.23",
    "langgraph>=1.1.0",
    "pydantic-settings>=2.13.1",
    "google-search-results>=2.4.2",
]
